+++ src/test/resources/com/pestphp/pest/PestUtil/PestItFunctionCallWithConcatString.php
@@ -1,0 +1,5 @@
+<?php
+
+it('basic ' . "super" . "test", function () {
+    $this->assertTrue(true);
+});