+++ src/test/resources/com/pestphp/pest/PestUtil/PestTestFunctionCallWithConcatString.php
@@ -1,0 +1,5 @@
+<?php
+
+test('basic ' . "su" . "per", function () {
+    $this->assertTrue(true);
+});